Tri des commentaires avec ajout des dates

+++ backend/controllers/publication.js
@@ -3,7 +3,7 @@ const fs = require('fs');
 const { connect } = require('../mysql-config');
 
 module.exports.getAllPublication = (req, res) => {
-    connectMysql.query('SELECT * FROM publication', (err, result) => {
+    connectMysql.query('SELECT * FROM publication ORDER BY id DESC', (err, result) => {
         if(err){
             console.log(err);
         } else {
@@ -23,19 +23,34 @@ module.exports.getOnePublication = (req, res) => {
 };
 
 module.exports.createPublication = (req, res) => {
+    var date = new Date();
+    var options = { year: "numeric", month: "long", day: "2-digit"};
+
     const newPublication = {
         userId: '55', //req.body.userId,
         textField: req.body.textField,
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        date: `${date.toLocaleDateString("fr-FR", options)} à ${('0'+date.getHours()).slice(-2)}:${('0'+date.getMinutes()).slice(-2)}`
     }
-    
-    connectMysql.query('INSERT INTO publication SET userId=?, textField=?, imageUrl=?', [newPublication.userId, newPublication.textField, newPublication.imageUrl],(err, result) => {
-        if(err){
-            console.log(err);
-        } else {
-            console.log("Publication créée");
-        }
-    })
+
+    if(req.file === undefined){
+        connectMysql.query('INSERT INTO publication SET userId=?, textField=?, date=?', [newPublication.userId, newPublication.textField, newPublication.date],(err, result) => {
+            if(err){
+                console.log(err);
+            } else {
+                console.log("Publication créée");
+            }
+        })
+    } else {
+        const imageUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`;
+
+        connectMysql.query('INSERT INTO publication SET userId=?, textField=?, imageUrl=?, date=?', [newPublication.userId, newPublication.textField, imageUrl, newPublication.date],(err, result) => {
+            if(err){
+                console.log(err);
+            } else {
+                console.log("Publication créée");
+            }
+        })
+    }
 };
 
 module.exports.modifyPublication = (req, res) => {
@@ -69,16 +84,34 @@ module.exports.deletePublication = (req, res) => {
             res.status(400).json({error : "La publication n'existe pas"})
         } else {
             const publication = result[0];
-            const filename = publication.imageUrl.split('/images/')[1]
-            fs.unlink(`images/${filename}`, () => {
+
+            if(publication.imageUrl == undefined){
                 connectMysql.query('DELETE FROM publication WHERE id=?', publication.id, (err, result) => {
-                    if(err) {
-                        res.status(400).json({error : "La publication n'existe pas"})
+                    if(err){
+                        console.log(err);
                     } else {
                         res.status(200).json({message: 'Publication Supprimée'});
                     }
                 })
-            })
+            } else {
+                const filename = publication.imageUrl.split('/images/')[1]
+                fs.unlink(`images/${filename}`, () => {
+                    connectMysql.query('DELETE FROM publication WHERE id=?', publication.id, (err, result) => {
+                        if(err) {
+                            res.status(400).json({error : "La publication n'existe pas"})
+                        } else {
+                            connectMysql.query('DELETE FROM comments WHERE publication_id=?', publication.id, (err, result) => {
+                                if(err){
+                                    console.log(err);
+                                } else {
+                                    console.log(result);
+                                }
+                                res.status(200).json({message: 'Publication Supprimée'});
+                            })
+                        }
+                    })
+                })
+            }
         }
     })
 };
